+++ src/test/java/com/arangodb/springframework/repository/query/derived/DerivedQueryCreatorTest.java
@@ -164,6 +164,41 @@ public class DerivedQueryCreatorTest extends AbstractArangoRepositoryTest {
 	}
 
 	@Test
+	public void findByStringContaining() {
+		final Customer c1 = new Customer("abc", "", 0);
+		final Customer c2 = new Customer("Abc", "", 0);
+		final Customer c3 = new Customer("abcd", "", 0);
+		final Customer c4 = new Customer("ab", "", 0);
+		repository.save(c1);
+		repository.save(c2);
+		repository.save(c3);
+		repository.save(c4);
+		final Collection<Customer> toBeRetrieved = new LinkedList<>();
+		toBeRetrieved.add(c1);
+		toBeRetrieved.add(c3);
+		final Collection<Customer> retrieved = repository.findByNameContaining("abc");
+		assertTrue(equals(retrieved, toBeRetrieved, cmp, eq, false));
+	}
+
+	@Test
+	public void findByStringContainingIgnoreCase() {
+		final Customer c1 = new Customer("abc", "", 0);
+		final Customer c2 = new Customer("Abc", "", 0);
+		final Customer c3 = new Customer("abcd", "", 0);
+		final Customer c4 = new Customer("ab", "", 0);
+		repository.save(c1);
+		repository.save(c2);
+		repository.save(c3);
+		repository.save(c4);
+		final Collection<Customer> toBeRetrieved = new LinkedList<>();
+		toBeRetrieved.add(c1);
+		toBeRetrieved.add(c2);
+		toBeRetrieved.add(c3);
+		final Collection<Customer> retrieved = repository.findByNameContainingIgnoreCase("abc");
+		assertTrue(equals(retrieved, toBeRetrieved, cmp, eq, false));
+	}
+
+	@Test
 	public void findTest() {
 		final Customer customer1 = new Customer("%_\\name", "%surname%", 20);
 		repository.save(customer1);